Fix for overflow, similar to NCBC-692

+++ Src/Couchbase/Configuration/Server/Serialization/Hdd.cs
@@ -14,7 +14,7 @@ namespace Couchbase.Configuration.Server.Serialization
         public long Used { get; set; }
 
         [JsonProperty("usedByData")]
-        public int UsedByData { get; set; }
+        public long UsedByData { get; set; }
 
         [JsonProperty("free")]
         public long Free { get; set; }
